@@ -36,8 +36,8 @@ module RailsBaseApi
     config.action_dispatch.default_headers = {}
     config.middleware.insert_before 0, "Rack::Cors" do
       allow do
-        origins '*'
-        resource '*', :headers => :any, :methods => [:get, :post, :options]
+        origins "*"
+        resource "*", headers: :any, methods: %i(get post options)
       end
     end
   end
