@@ -1,6 +1,5 @@
-public_key_path = "#{Rails.root}/rsa_public.pem"
+public_key_path = Rails.root.join("rsa_public.pem")
 RSA_PUBLIC_KEY = OpenSSL::PKey::RSA.new(File.read(public_key_path))
-private_key_path = "#{Rails.root}/rsa_private.pem"
+private_key_path = Rails.root.join("rsa_private.pem")
 password = ENV["RSA_PASSWORD"]
 RSA_PRIVATE_KEY = OpenSSL::PKey::RSA.new(File.read(private_key_path), password)
-
