@@ -1,4 +1,4 @@
-SimpleTokenAuthentication.configure do |config| # rubocop:disable Lint/UnusedBlockArgument
+SimpleTokenAuthentication.configure do |config|
   # Configure the session persistence policy after a successful sign in,
   # in other words, if the authentication token acts as a signin token.
   # If true, user is stored in the session and the authentication token and
