@@ -1,4 +1,6 @@
 FactoryGirl.define do
-  sequence(:email) { |n| "user#{n}@example.com" }
+  sequence(:username) { |n| "user#{n}" }
   sequence(:password) { "123456" }
+  sequence(:article_title) { Faker::Lorem.sentence }
+  sequence(:article_content) { Faker::Lorem.paragraph }
 end
