@@ -1,8 +1,8 @@
 require "rails_helper"
 
 describe UserSerializer do
-  let(:user) { build :user, id: 1, authentication_token: "token" }
-  let(:json) { ActiveModel::SerializableResource.serialize(user).to_json }
+  let(:user) { build :user }
+  let(:json) { ActiveModelSerializers::SerializableResource.new(user).to_json }
   let(:user_json) { parse_json(json)["user"] }
 
   it "returns user" do
