@@ -1,9 +1,8 @@
-RSpec::Matchers.define :be_a_session_representation do
+RSpec::Matchers.define :be_a_token_representation do
   match do |json|
     response_attributes = %w(
-      id
-      authentication_token
-      email
+      auth_token
+      refresh_token
     )
 
     expect(json).to be
