@@ -1,8 +1,9 @@
 RSpec::Matchers.define :be_a_user_representation do |user|
   match do |json|
-    response_attributes = user.sliced_attributes %w(
+    response_attributes = user.attributes.slice %w(
       id
-      email
+      username
+      role
     )
 
     expect(json).to be
